+++ CHANGELOG.md
@@ -11,6 +11,20 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ## [Released]
 
+## [2.2.0] - 2020-11-11
+
+### Added
+- error notifications for invalid config options
+
+### Changed
+- some internal methods for more performance
+
+### Fixed
+- bot invoker logic
+- duplicated entries error when the bot assigns/removes groups
+- group removal logic
+
+
 ## [2.1.1] - 2020-10-26
 
 ### Fixed
@@ -37,6 +51,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 <!-- Versions -->
 [Unreleased]: https://github.com/RLNT/sinus-automated-servergroups/compare/v2.0.0...HEAD
 [Released]: https://github.com/RLNT/sinus-automated-servergroups/releases
+[2.2.0]: https://github.com/RLNT/sinus-automated-servergroups/compare/v2.1.1..v2.2.0
 [2.1.1]: https://github.com/RLNT/sinus-automated-servergroups/compare/v2.1.0..v2.1.1
 [2.1.0]: https://github.com/RLNT/sinus-automated-servergroups/compare/v2.0.0..v2.1.0
 [2.0.0]: https://github.com/RLNT/sinus-automated-servergroups/releases/v2.0.0